chore: move interfaces to types folder
